// app/plantilla2/src/app/demo/models/GraficosViewModel.ts
export class Graficos {
    usua_Usuario?: string;
    comb_Id?: number;
    comb_Descripcion?: string;
    totalPedidosCombos?: number;

    post_id?: number;
    post_Descripcion?: string;
    totalPedidosPostres?: number;

    paqe_Id?: number;
    paqe_Descripcion?: string;
    totalPedidosPaquetes?: number;

    alim_Id?: number;
    alim_Descripcion?: string;
    totalPedidosAlimentos?: number;
}